changed default feedforward model to pt
